[MODFISTO-393] - refactoring
