+++ src/test/java/org/folio/consortia/context/PostgresTestContainersBeanFactoryPostProcessor.java
@@ -1,4 +1,4 @@
-package org.folio.template.context;
+package org.folio.consortia.context;
 
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanFactoryPostProcessor;
